added diffrent component for purchase,dashboard working on purchase,used alerts

+++ Frontend/src/app/purchase/components/products-list/products-list.component.ts
@@ -12,31 +12,42 @@ import { PurchaseService } from '../../services/purchase.service';
 })
 
 export class ProductsListComponent implements OnInit {
-  @ViewChild(MatPaginator) paginator!: MatPaginator;
+  @ViewChild(MatPaginator) set paginator(matPaginator:MatPaginator){
+    if(matPaginator!=undefined){
+      this.purchaseList.paginator=matPaginator;
+    }
+  }
 
   constructor(public sharedService: SharedService,private purchaseService:PurchaseService,private router:Router) { }
-  purchaseList!: MatTableDataSource<any>;
-  displayedColumns = ["vehicle_no", "seller_name", "purchase_date", "actions"]
+  purchaseList!:MatTableDataSource<any>;
+  length:any;
+
+  displayedColumns = ["vehicle_no","model","car_name" ,"seller_name", "purchase_date", "actions"]
   ngOnInit(): void {
+    this.purchaseService.allDetails=this.purchaseService.vehicleDetails="";
+
     this.sharedService.viewPurchaseDetails().subscribe(res=>{
       if(res){
         this.purchaseList=new MatTableDataSource(res.data);
-      }
-      else{
-        alert('Something went wrong')
+        this.length=this.purchaseList.data.length
       }
     })
   }
 
   onPurchaseDelete(carNo:any){
-    this.purchaseService.deletePurchase(carNo).subscribe((res)=>{
-      if(res){
-        alert("Vehicle "+carNo+" "+res.message)
-        this.ngOnInit();
-      }else{
-        alert(res.message)
-      }
-    })
+    if(confirm(`Are You Sure to Delete Vehicle ${carNo}`)){
+      this.purchaseService.deletePurchase(carNo).subscribe((res)=>{
+        if(res){
+          alert("Vehicle "+carNo+" "+res.message)
+          this.ngOnInit();
+        }else{
+          alert(res.message)
+        }
+      })
+    }else{
+      return
+    }
+
    }
 
    onPurchaseAdd(){
@@ -46,4 +57,10 @@ export class ProductsListComponent implements OnInit {
    onUpdatePurchase(carNo:any){
     this.router.navigate(['admin/purchase/addvehicle'],{queryParams:{carno:carNo}})
    }
+
+   applyFilter(event:any){
+    const filterValue=event.target.value;
+    this.purchaseList.filter=filterValue.trim().toLowerCase();
+   }
+
 }
